changed conv layer to size_t

+++ CNN_Test/convolutional_layer_test.cpp
@@ -12,9 +12,9 @@ namespace CNNTest
 		{
 			convolutional_layer layer(3, 2, 1, e_activation_t::relu_fn);
 			Assert::AreEqual((int)e_layer_type_t::convolution, (int)layer.get_layer_type());
-			Assert::AreEqual(3, (int)layer.get_kernel_count());
-			Assert::AreEqual(2, layer.get_kernel_size());
-			Assert::AreEqual(1, layer.get_stride());
+			Assert::AreEqual((size_t)3, layer.get_kernel_count());
+			Assert::AreEqual((size_t)2, layer.get_kernel_size());
+			Assert::AreEqual((size_t)1, layer.get_stride());
 		}
 		TEST_METHOD(invalid_constructor)
 		{
